requirements.txt
pandas==2.2.3
scikit-learn==1.5.2
datasets==3.1.0
PyYAML==6.0.2
transformers==4.46.3
torch==2.10.0
mlflow==2.18.0
dvc==3.58.0
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.9.2
streamlit==1.40.2
requests==2.32.3
python-dotenv==1.0.1
huggingface_hub==0.26.2
openpyxl==3.1.5
pytest==8.3.3
httpx==0.27.2
